FMT: yet more thrashing
With black 20.8b0 if there is a trailing comment then black will
explode to the one-entry-per-line mode rather than bring to one
line (even if it would fit).

In black 19.10b1 it would re-format those statements to one line (but
not strip the trailing ,).

Because we went through a cycle of shorter line length formatting ->
longer line length formatting we ended up with commas from the
multi-line explosion that then got "locked in".