FIX: reject create/close environment requests if already underway
Make use of the RE Manager MState enum to indicate when the
manager is already underway in the creation or closing of a
RE Worker environment.

Previously the manager would get into a locked up state if
a request to create an environment was received while it was
already mid-way through the act of creating an environment.