ENH: stop the queue if outstanding unresolved pause request
Previous behaviour: if a plan is currently being executed and then the user
issues a request to pause (with option deferred) the RE worker will pause the
plan at the next checkpoint in the plan, if however the plan has already been
executed beyond the last checkpoint in the plan, then the plan finishes
completely and the queueserver then sees that the plan finished, so serves the
worker the next plan in the queue.

Expectation: A human user issuing the command to pause at the next checkpoint
would likely expect the queue execution to come to a stop at some point in the
near future, it could be argued that they would not expect the plan that they
told to pause (deferred) to finish and then for the queueserver to carry on
executing plans, because that feels like their request to pause, which was
successfully accepted, was actually just completely ignored.
An argument could be made that the user should have issued a stop queue command
if that's what they wanted, but this argument relies on the user knowing that
the final checkpoint had already been passed, which I think is unrealistic.

Current behaviour: If a request to pause is successfully accepted, the
queueserver will EITHER pause at the next checkpoint in the current plan, OR
stop the queue once the current plan finishes, whichever comes first.